fix(Build): Fix authOptions build error
