+++ src/app/api/session/route.ts
@@ -10,7 +10,7 @@
 
 import { getServerSession } from 'next-auth'
 import { NextResponse } from 'next/server'
-import { authOptions } from '../auth/[...nextauth]/route'
+import { authOptions } from '../auth/[...nextauth]/authOptions'
 
 export async function GET() {
     const session = await getServerSession(authOptions)
